Created submission_functions.ts and supabaseClient.ts for functions, not tested yet.
