+++ src/lib/supabaseClient.ts
@@ -1,0 +1,5 @@
+import { createClient } from '@supabase/supabase-js';
+
+const supabaseURL = import.meta.env.VITE_SUPABASE_URL;
+const supabaseAnonKey = import.meta.env.VITE_SUPABASE_ANON_KEY;
+export const supabaseClient = createClient(supabaseURL, supabaseAnonKey);
