+++ src/routes/member_submissions/+page.svelte
@@ -3,7 +3,9 @@
     import '../../app.postcss';
     import { onMount } from 'svelte';    
     import { goto } from '$app/navigation'; 
-    
+    import { selectCategory } from '$lib/submission_functions';
+    import { handleSubmit } from '$lib/submission_functions';  
+        
     const supabaseURL = import.meta.env.VITE_SUPABASE_URL;
     const supabaseAnonKey = import.meta.env.VITE_SUPABASE_ANON_KEY;
     const supabaseClient = createClient(supabaseURL, supabaseAnonKey);
@@ -16,9 +18,7 @@
     let r_recipes_preparation_time: number = 0;
     let r_recipes_cooking_time: number = 0;
     let r_recipes_servings: number = 0;
-    let selectedCategory: number | null = null;    
-    let imageFile: File | null = null;   
-
+    let selectedCategory: number | null = null;        
     
     onMount(async () => {
         const { data: categoriesData, error: categoriesError } = await supabaseClient
@@ -34,100 +34,11 @@
                 return map;
             }, {});
         }
-    });  
-    
-    async function uploadImage() {
-        if (imageFile) {
-            const fileName = `${Date.now()}-${imageFile.name}`;
-            const { data: uploadData, error: uploadError } = await supabaseClient.storage
-                .from('recipes_images')
-                .upload(fileName, imageFile);
-
-            if (uploadError) {
-                console.error('Error uploading image:', uploadError.message);
-                return null;
-            }
-
-            const { data: publicUrlData } = supabaseClient.storage
-                .from('recipes_images')
-                .getPublicUrl(fileName);
-
-            return publicUrlData.publicUrl; 
-        }
-        return null;
-    }
-
-    async function createRecipe(imageUrl: string | null) {
-        const { data, error } = await supabaseClient
-            .from('Recipes')
-            .insert([{
-                r_recipes_title,
-                r_recipes_description,
-                r_recipes_instructions,
-                r_recipes_preparation_time,
-                r_recipes_cooking_time,
-                r_recipes_servings,
-                c_category_id: selectedCategory,
-                r_recipes_image: imageUrl || null
-            }])
-            .single();
-
-        if (error) {
-            console.error('Error inserting recipe:', error.message);
-            return null;
-        } else {
-            const recipeId = (data as { id: number })?.id;
-            console.log('Recipe created with ID:', recipeId);          
-            
-            if (recipeId) {
-                const { data: recipeData, error: recipeError } = await supabaseClient
-                    .from('Recipes')
-                    .select('*')
-                    .eq('r_recipes_id', recipeId)
-                    .single();
-
-                if (recipeError) {
-                    console.error('Error fetching recipe:', recipeError);
-                } else {
-                    console.log('Recipe fetched:', recipeData);
-                }                            
-              }
-            // Clear form fields
-            r_recipes_title = '';
-            r_recipes_description = '';
-            r_recipes_instructions = '';
-            r_recipes_preparation_time = 0;
-            r_recipes_cooking_time = 0;
-            r_recipes_servings = 0;
-            selectedCategory = null;
-            imageFile = null;
-            
-            return recipeId;
-        }
-    }
-
-    function selectCategory(categoryId: number) {
-        selectedCategory = categoryId;
-    }
-    
-    async function handleSubmit() {
-        const imageUrl = imageFile ? await uploadImage() : null;
-        const recipeId = await createRecipe(imageUrl);
-        
-        if (recipeId) {
-            goto('/member_submissions/recipe_ingredients');
-        } else {
-            console.error('Recipe creation failed.');
-        }
-    }
-
-    function handleFileChange(event: Event) {
-        const target = event.target as HTMLInputElement;
-        imageFile = target.files ? target.files[0] : null;
-    }
+    }); 
+              
 </script>
 
-<h1>Recipes</h1>
+<h1 class="heading">Recipes</h1>
 
 <form on:submit|preventDefault={handleSubmit}>
     <table>
@@ -182,6 +93,6 @@
             </td>
         </tr>
     </table>
-
+    <button id="cr" type="submit">Create Recipe & Add Details</button>      
     <button id="cr" type="button" on:click={() => goto('/member/logged_in')}>CANCEL</button>
 </form>
